tests/wpt/web-platform-tests/speculation-rules/prerender/restriction-web-serial.tentative.https.html

<!DOCTYPE html>
<title>Access to the Web Serial API is deferred</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const uid = token();
  const bc = new PrerenderChannel('test-channel', uid);

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  const url = `resources/web-serial.https.html?uid=${uid}`;
  window.open(url, '_blank', 'noopener');

  const result = await gotMessage;
  const expected = [
    {event: 'started waiting navigator.serial.getPorts',  prerendering: true},
    {event: 'prerendering change',                        prerendering: false},
    {event: 'finished waiting navigator.serial.getPorts', prerendering: false},
  ];
  // The spec, https://wicg.github.io/nav-speculation/prerendering.html#patch-serial,
  // says that Web Serial API has delay while prerendering for
  // requestPort(). As this test uses getPorts(), it's a tentative test.
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event[${i}]`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering[${i}]`);
  }
  bc.close();

  // Send a close signal to PrerenderEventCollector on the prerendered page.
  new PrerenderChannel('close', uid).postMessage('');
}, `the access to the Web Serial API should be deferred until the prerendered
    page is activated`);

</script>
</body>
